Relocate SmithingListener into a separate package
